requirements.txt
#decrypt requirements
azure-mgmt-resource>=18.0.0
azure-identity>=1.5.0
azure-keyvault-secrets
azure.keyvault.keys